requirements.txt
azure.storage.blob==12.25.1
beanie==1.29.0
beautifulsoup4==4.13.4
camelot-py[cv]
celery[redis]
celery==5.5.2
duckduckgo_search==8.0.1
fastapi==0.115.12
google-auth-oauthlib==1.2.2
google-auth==2.40.0
httpx==0.28.1
ipinfo==5.1.1
ipython==8.36.0
janus==2.0.0
langchain-community==0.3.23
langchain-core==0.3.58
langchain-google-genai==2.1.4
langchain-openai==0.3.16
langchain-tavily==0.1.6
langchain==0.3.25
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langgraph==0.4.1
litellm==1.61.0
lxml==5.4.0
markdown-it-py==3.0.0
markdown==3.6
matplotlib==3.10.1
mdit-py-plugins==0.4.2
motor==3.7.0
numpy==2.2.5
openai==1.75.0
pandas==2.2.3
passlib[bcrypt]==1.7.4
pdfkit==1.0.0
pdfplumber==0.11.7
plotly==6.0.1
pydantic[email]==2.11.4
pydantic==2.11.4
Pygments==2.19.1
pymongo==4.12.1
PyMuPDF==1.25.5
pypdf==5.4.0
PyPDF2==3.0.1
python-docx==1.1.2
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
qdrant-client==1.14.2
scikit-learn==1.6.1
seaborn==0.13.2
selenium==4.32.0
statsmodels==0.14.4
tldextract==5.3.0
tzlocal==5.3.1
uvicorn==0.34.2
weasyprint==62.3
webdriver-manager==4.0.2
yahooquery==2.4.1
yfinance==0.2.61